Moved modular GUI to differentiate easier
