+++ src/tracker/Pages/ModularGUI/PromptsPage.java
@@ -1,0 +1,4 @@
+package tracker.Pages.ModularGUI;
+
+public class PromptsPage {
+}
