+++ src/tracker/Pages/ModularGUI/SettingsPage.java
@@ -1,0 +1,4 @@
+package tracker.Pages.ModularGUI;
+
+public class SettingsPage {
+}
