+++ src/tracker/Pages/ModularGUI/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Modular GUI pages generated using GUI components
+ */
+package tracker.Pages.ModularGUI;
