--- src/tracker/Pages/PromptsPage.java
@@ -1,4 +1,0 @@
-package tracker.Pages;
-
-public class PromptsPage {
-}
