--- src/tracker/Pages/SettingsPage.java
@@ -1,4 +1,0 @@
-package tracker.Pages;
-
-public class SettingsPage {
-}
